refactor(table-pagination): moved action function to separate file
