+++ app/client/src/features/admin/UsersTable.tsx
@@ -12,73 +12,15 @@ import {
     TextField,
     IconButton,
     TableFooter,
-    Box,
     TablePagination,
 } from '@mui/material';
-import { useTheme } from '@mui/material/styles';
 import { Search as SearchIcon } from '@mui/icons-material';
-import FirstPageIcon from '@mui/icons-material/FirstPage';
-import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import LastPageIcon from '@mui/icons-material/LastPage';
 
 import { Form } from '@local/components';
 import type { useUsersDashboardFragment$key } from '@local/__generated__/useUsersDashboardFragment.graphql';
 import { useUsersDashboard } from './useUsersDashboard';
 import { useForm } from '@local/core';
-
-interface TablePaginationActionsProps {
-    count: number;
-    page: number;
-    rowsPerPage: number;
-    onPageChange: (event: React.MouseEvent<HTMLButtonElement>, newPage: number) => void;
-}
-
-function TablePaginationActions(props: TablePaginationActionsProps) {
-    const theme = useTheme();
-    const { count, page, rowsPerPage, onPageChange } = props;
-
-    const handleFirstPageButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        onPageChange(event, 0);
-    };
-
-    const handleBackButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        onPageChange(event, page - 1);
-    };
-
-    const handleNextButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        onPageChange(event, page + 1);
-    };
-
-    const handleLastPageButtonClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        onPageChange(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-    };
-
-    return (
-        <Box sx={{ flexShrink: 0, ml: 2.5 }}>
-            <IconButton onClick={handleFirstPageButtonClick} disabled={page === 0} aria-label='first page'>
-                {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-            </IconButton>
-            <IconButton onClick={handleBackButtonClick} disabled={page === 0} aria-label='previous page'>
-                {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-            </IconButton>
-            <IconButton
-                onClick={handleNextButtonClick}
-                disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                aria-label='next page'
-            >
-                {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-            </IconButton>
-            <IconButton
-                onClick={handleLastPageButtonClick}
-                disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                aria-label='last page'
-            >
-                {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-            </IconButton>
-        </Box>
-    );
-}
+import { TablePaginationActions } from './TablePaginationActions';
 
 type UsersDashboardSearchFilter = {
     firstName: string;
@@ -158,7 +100,7 @@ export function UsersTable({ fragmentRef }: UsersTableProps) {
     const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
     const date = new Date();
-    const FETCH_AMMOUNT = 50;
+    const FETCH_AMMOUNT = 100;
 
     // Avoid a layout jump when reaching the last page with empty rows.
     const emptyRows = React.useMemo(
@@ -191,8 +133,8 @@ export function UsersTable({ fragmentRef }: UsersTableProps) {
     };
 
     const handleLoadNext = React.useCallback(() => {
-        if (hasNext) loadNext(FETCH_AMMOUNT);
-    }, [hasNext, loadNext]);
+        if (hasNext && !isLoadingNext) loadNext(FETCH_AMMOUNT);
+    }, [hasNext, loadNext, isLoadingNext]);
 
     const usersListLength = React.useMemo(() => users.length, [users]);
     const nextPageIsLastPage = React.useMemo(
@@ -302,15 +244,6 @@ export function UsersTable({ fragmentRef }: UsersTableProps) {
                 <Grid item container justifyContent='center' paddingY='2rem'>
                     <Typography>No users found</Typography>
                 </Grid>
-            )}
-            {hasNext ? (
-                <Grid container justifyContent='center' paddingY='2rem'>
-                    <Button variant='contained' disabled={isLoadingNext} onClick={handleLoadNext}>
-                        Load More
-                    </Button>
-                </Grid>
-            ) : (
-                <React.Fragment />
             )}
         </React.Fragment>
     );
